requirements.txt
pandas
requests
numpy
tqdm
tenacity
langchain
openai
python-dotenv
argparse

beautifulsoup4
lxml
django>=5.0
djangorestframework>=3.15
django-filter>=24.2
dj-database-url>=2.2
psycopg2-binary>=2.9
sentry-sdk>=2.0
django-cors-headers>=4.3
gunicorn>=22.0
